Refactor: extract guide page module into web/js/pages/guide.js (#78)
